Split data-management.tsx into TierPanel + OrgAccountsSection
Pure refactor — no behavior change. The 954-line view is now three
files in views/, each focused on one responsibility:

- data-management-tier.tsx (TierPanel + formatBytes/formatPeriod helpers
  + SyncState type) — the per-tier panel rendered three times for
  daily, hourly, cost-optimization.
- data-management-org.tsx (OrgAccountsSection) — the AWS Organizations
  sync UI section.
- data-management.tsx — the DataManagement view orchestrating tier
  panels and the org section.
